Adds (back) GCC compatibility. Removes test/ folder. Note: CMake scripts are forthcoming. This commit is source code only. There's new precompiled headers. The test/ folder will not be built by the new scripts, and is deleted only because it's more checkboxes than I care to uncheck.
git-svn-id: https://svn.code.sf.net/p/collada-dom/code/trunk@917 dfea1ed5-6b0d-0410-973e-e49e28f48b26

+++ dom/src/dae/domAny.cpp
@@ -60,7 +60,7 @@ domAny::~domAny()
 	//a shared metadata object.	
 	_meta->_attribs.~daeArray(); 
 	//delete/destruct the attribute-buckets.
-	_attribs._self_destruct(_meta->_attribs.size());
+	_attribs._self_destruct((int)_meta->_attribs.size());
 }
 void domAny::AttributeBucket::_self_destruct(int i)
 {
